requirements.txt
Django==3.0.4
python-decouple==3.1
psycopg2==2.8.5
django-crontab==0.7.1
djangorestframework==3.11.0
django-environ==0.9.0
requests==2.23.0